Add files via upload
The purpose of this project is to create a Python program for converting numbers between different numerical systems along with implementing three arithmetic operations(addition, subtraction, and multiplication) on binary numbers. The program will support the following numerical systems: binary, ternary, quaternary, octal, decimal, duodecimal, hexadecimal, and vigesimal. The program will be designed to take input from the user in any of these numerical systems and convert it to any other supported system and print out the result using a command-line interface.